requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
PyJWT==2.8.0
google-generativeai==0.3.2
openpyxl==3.1.2
python-dotenv==1.0.0
